requirements.txt
torch
pytorch-lightning
pandas
transformers
transformers-interpret
streamlit
click
hydra-core
omegaconf